requirement.txt
selenium>=4.10
beautifulsoup4>=4.12
sqlalchemy>=2.0
celery>=5.3
redis>=4.5
pandas>=2.0
python-dotenv>=1.0
requests>=2.30